+++ frontend/src/lib/utils/loader/ProcessFile.ts
@@ -1,5 +1,4 @@
 import Worker from '$lib/utils/workers/FileProcessWorker.ts?worker';
-import { on } from 'svelte/events';
 
 export const processFile = async (
   files: FileList,
@@ -10,54 +9,58 @@ export const processFile = async (
     onPreview,
     onFull,
     onProgress,
+    onError,
   }: {
-    onPreview: (data: { previewData: any[] }) => void,
-    onFull: (data: { fileData: any[], fileName: string, fileType: string, fileSize: number }) => void
+    onPreview: (data: { previewData: any[]; metadata: any }) => void,
+    onFull: (data: { fileData: any[], metadata: any }) => void
     onProgress: (data: { message: string, percentage: number }) => void,
+    onError: (data: { error: string }) => void,
   }
 ) => {
   if (!files?.[0]) return;
   const file = files[0];
 
+  const worker = new Worker(); 
+  worker.onmessage = (event: MessageEvent) => {
+    const { type, ...data } = event.data;
 
-  return new Promise<void>((resolve, reject) => {
-    const worker = new Worker();
-    console.log("Worker initialized for file processing");
-    worker.onmessage = (e) => {
-      const { type, error, previewData, fileData } = e.data;
-
-      if (error) {
+    switch (type) {
+      case 'preview':
+        onPreview(data as { previewData: any[]; metadata: any });
+        break;
+      case 'full':
+        onFull(data as { fileData: any[], metadata: any });
         worker.terminate();
-        reject(new Error(error));
-        return;
-      }
-      if (type === "progress") {
-        onProgress({ message: e.data.message, percentage: e.data.percentage });
-        console.log("Progress update:", e.data.message, "Percentage:", e.data.percentage);
-      } else if (type === "preview") {
-        onPreview({ previewData });
-        console.log("Preview data received:", previewData.length, "rows");
-      } else if (type === "full") {
-        onFull({
-          fileData,
-          fileName: file.name,
-          fileType: file.type,
-          fileSize: file.size,
-        });
-        console.log("Full data received:", fileData.length, "rows");
-        console.log(fileData.slice(0, 5)); // Log first 5 rows for debugging
-        onProgress({ message: "Файл обролено", percentage: 100 });
+        break;
+      case 'progress':
+        onProgress(data as { message: string, percentage: number });
+        break;
+      case 'error':
+        onError(data as { error: string });
         worker.terminate();
-        resolve();
-      }
-    };
+        break;
+      default:
+        console.warn('Unknown message type from worker:', type, data);
+    }
+  };
 
-    worker.onerror = (err) => {
-      worker.terminate();
-      reject(err);
-    };
+  worker.onerror = (error) => {
+    console.error("Worker error:", error);
+    onError({ error: `Помилка Web Worker: ${error.message || 'Невідома помилка'}` });
+  };
 
-    // ✅ передаємо startFrom
-    worker.postMessage({ file: file, fileName: file.name, startFrom, previewRows, startPreviewFrom });
-  });
+  try {
+    const fileBuffer = await file.arrayBuffer();
+    worker.postMessage({
+      fileBuffer,
+      fileName: file.name,
+      startFrom,
+      previewRowsCount: previewRows,
+      startPreviewFrom,
+    });
+  } catch (error: any) {
+    console.error("Error reading file:", error);
+    onError({ error: `Помилка читання файлу: ${error.message || 'Невідома помилка'}` });
+  }
 };
+
