+++ frontend/src/lib/utils/loader/TransformFile.svelte.ts
@@ -1,5 +1,30 @@
-import type { MappedHeader, TransformedItem, WorkerMessage } from "../workers/TransfromFileWorker";
 import Worker from '$lib/utils/workers/TransfromFileWorker.ts?worker';
+
+export type MappedHeader = {
+  name: string;
+  value: string;
+  type: string;
+  header: string;
+  aliases?: string[];
+};
+
+export interface TransformedItem {
+  brand: string;
+  article: string;
+  price: number;
+  description: string;
+  provider_id: string;
+  rests: {
+    [warehouseId: string]: number | string;
+  };
+}
+
+export type WorkerMessage =
+  | { type: 'startUpload'; data: { data: any[]; mappedHeaders: MappedHeader[]; providerId: string; companyId: string; authToken: string; } }
+  | { type: 'progress'; payload: { state: "transforming" | "hash" }; }
+  | { type: 'complete'; payload: { transformedData: TransformedItem[]; hash: string }; }
+  | { type: 'error'; payload: { message: string }; };
+
 
 export async function StartTransformFileWorker(
   fileData: any[],
